First working commit
In this commit, the program now effectively can download a day of FPDS data
and convert it into a single json file, by running the command:

`python run.py <date>`

where date is in the format "YYYY/MM/DD".

The file will save in a `results/` folder with the `<date>.json` as the name of the
file.