Instead of saving to memory, write to disk
Although i/o operations are slow, keeping all of the data saved in a giant
array seems like a recipe for disaster. Instead, just write to disk.

A still better improvement would be to save to a buffer and then save to
disk in chunks, but that's optimization I'm not doing today.